Add a demo computing the SIC at .1% background efficiency
The notebook produces a single point of Figs. 5 and 6 in the Back to the
Roots paper [2309.13111]. The train/val/test split is different so the
numbers differ slightly without contradicting the paper.

The running time of this notebook is unfortunately quite long due to the
ensembling of 100 tree models, so I recorded my numbers for different
amounts of signal in the last cell.